Вывод кнопок в зависимости от роли

+++ resources/views/news/show.blade.php
@@ -4,11 +4,17 @@
     <div class="container">
         <div class="row">
             <div class="col-md-10"><h1>{{ $news->title }}</h1></div>
-            <div class="col-md-1"><a class="btn btn-warning" href="{{ route('news.edit', $news->slug) }}">Edit</a></div>
             <div class="col-md-1">
-                {!! Form::open(['route' => ['news.destroy', $news->slug], 'method' => 'DELETE']) !!}
-                <button class="btn btn-danger">Del</button>
-                {!! Form::close() !!}
+                @if($roles->contains('moderator'))
+                    <a class="btn btn-warning" href="{{ route('news.edit', $news->slug) }}">Edit</a>
+                @endif
+            </div>
+            <div class="col-md-1">
+                @if($roles->contains('admin'))
+                    {!! Form::open(['route' => ['news.destroy', $news->slug], 'method' => 'DELETE']) !!}
+                    <button class="btn btn-danger">Del</button>
+                    {!! Form::close() !!}
+                @endif
             </div>
         </div>
         <div class="row">
